Add scripts to test minimal installation mode of EFA Installer
Since 1.8.0, EFA Installer has a new flag "-m" which run installer
in minimal mode. Under this mode, only kernel module and rdma-core
are installed, libfabric and openmpi are not installed.

This mode is intended for users who only want to use Intel MPI
(2019U6 and above), because Intel MPI 2019U6 comes with a copy
of libfabric.

Because the libfabric comes with Intel MPI does not contain headers,
fabtests cannot be built and run. Therefore, this set of script will
install Intel MPI 2019U6 and run osu MPI test using it.

Signed-off-by: Wei Zhang <wzam@amazon.com>